edició migracions, seeders i correccions en forms de creació/edició amb valors old() si no es valida correctament

+++ database/migrations/2017_04_06_183650_create_productes_table.php
@@ -20,7 +20,7 @@ class CreateProductesTable extends Migration
 			$table->integer('list_id')->unsigned();
 			$table->boolean('confirmed')->default(0);
 			$table->string('name')->required();
-			$table->integer('quantity')->nullable();
+			$table->decimal('quantity', 5, 2)->nullable();
 			$table->string('units')->nullable();
 			$table->decimal('price', 5, 2)->nullable();
 			$table->integer('assigned_to')->nullable()->unsigned();
